Service code/bbc/requirements.txt
pika 
newsapi-python
datetime
